[#46] Support expand whisper
